Replace disk->disk group reference with copied name
During DM table generate we need to generate canonical device names for both
volumes and disks which include the disk group name. Ideally we would reference
the disk group to do this, however this is greatly complicated by gobject's
reference counting, as it introduces circular references.

To get round this, we simply copy the DG data we need into the relevant
structures.