Remove Component from exposed API
Although the internal LDM data structure uses disk groups, volumes, components,
partitions and disks, components actually serve no useful purpose and only make
the data more unwieldy. This change removes components from the exposed API.
This makes it much simpler to understand how a volume is constructed.

All partitions of a component are moved direcly into the parent volume, and
pre-sorted by index. This makes it unnecessary to expose 'index' through the
api, so it is removed.

Windows exposes 5 types of volume:
* simple
* spanned
* striped
* mirrored
* raid5

However, these types don't directly exist in the underlying data due to the
confusing interaction of volumes and components. With the removal of components,
we now expose the type the user expects to see.

This change also renames private structures for brevity.