Cast result of be64toh() to uint64_t
glibc 2.16 makes a change to be64toh() which breaks libldm. In the new glibc,
be64toh() returns unsigned long long int regardless of wordsize. This breaks in
combination with PRIu64 on x86_64, which expects unsigned long.

It's not clear to me whether this is a regression in glibc, or a failure of gcc
to recognise that unsigned long long int and unsigned long are the same thing on
x86_64.

There's a Fedora bug open against glibc here:
https://bugzilla.redhat.com/show_bug.cgi?id=859427